test: Move inline to freezer
